requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
transformers>=4.30.0

# Build tools
maturin>=1.0.0

# Optional: for better performance on Apple Silicon
# accelerate>=0.20.0
